Create GitSourceFactory
